Create Linode TS namespace. Extract types.
