+++ src/components/TopMenu.tsx
@@ -68,4 +68,4 @@ class TopMenu extends React.Component<Props> {
 
 export default withStyles(styles, { withTheme: true })<Props>(
   TopMenu,
-) as TodoAny;
+) as Linode.TodoAny;
