+++ src/types/Image.ts
@@ -1,0 +1,16 @@
+namespace Linode {
+  export interface Image {
+    id: string;
+    label: string;
+    description: string | null;
+    created: string;
+    type: ImageType;
+    is_public: Boolean;
+    size: number;
+    created_by: string;
+    vendor: string;
+    deprecated: Boolean;
+  }
+
+  type ImageType = 'manual' | 'automatic';
+}
