+++ src/types/LinodeType.ts
@@ -1,0 +1,18 @@
+namespace Linode {
+  export interface LinodeType {
+    id: string;
+    disk: number;
+    class: LinodeTypeClass;
+    price: PriceObject;
+    label: string;
+    addons: {
+      backups: { price: PriceObject };
+    };
+    network_out: number;
+    memory: number;
+    transfer: number;
+    vcpus: number;
+  }
+
+  type LinodeTypeClass = 'nanode' | 'standard' | 'highmem';
+}
